requirements.txt
# Core
numpy==1.25.2
opencv-python==4.8.1.78
pillow==10.0.0
matplotlib==3.8.0

# PyTorch (YOLOv5 training / export)
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# ONNX
onnx==1.15.0
onnxruntime==1.16.1

# For YAML config parsing
pyyaml==6.0

# For CLI and utilities
tqdm==4.66.1
scipy==1.12.1

# Optional: if using pandas for any data handling
pandas==2.1.1

# Optional: for logging/debugging
loguru==0.7.0

# SNPE SDK is not pip-installable; must be installed separately
# See: https://developer.qualcomm.com/software/snpe
